pyproject.toml
[project]
name = "furniture"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.12.0",
    "dotenv>=0.9.9",
    "fastapi>=0.135.3",
    "google-genai>=1.70.0",
    "jupyterlab>=4.5.6",
    "nest-asyncio>=1.6.0",
    "pillow>=12.2.0",
    "playwright>=1.58.0",
    "playwright-stealth>=2.0.3",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.76.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "supabase>=2.28.3",
    "uvicorn[standard]>=0.44.0",
]
